+++ include/H2DE/engine/H2DE_lerp_manager.h
@@ -6,18 +6,6 @@ class H2DE_Engine;
 
 class H2DE_LerpManager {
 public:
-    template<typename H2DE_Lerp_T>
-    static H2DE_TimelineID lerp(H2DE_Engine* engine, H2DE_Lerp_T origin, H2DE_Lerp_T destination, H2DE_TimelineID duration, H2DE_Easing easing, const std::function<void(H2DE_Lerp_T)>& update, const std::function<void()>& completed, bool pauseSensitive = true) {
-        const H2DE_Lerp_T valueToAdd = destination - origin;
-
-        return engine->createTimeline(duration, easing, [update, origin, valueToAdd](float blend) {
-            if (update) {
-                H2DE_Lerp_T interpolatedValue = origin + (valueToAdd * blend);
-                update(interpolatedValue);
-            }
-        }, completed, 0, pauseSensitive);
-    }
-
     static H2DE_TimelineID lerp(H2DE_Engine* engine, const H2DE_ColorRGB& origin, const H2DE_ColorRGB& destination, H2DE_TimelineID duration, H2DE_Easing easing, const std::function<void(H2DE_ColorRGB)>& update, const std::function<void()>& completed, bool pauseSensitive = true) {
         const uint8_t redToAdd = destination.r - origin.r;
         const uint8_t greenToAdd = destination.g - origin.g;
@@ -33,6 +21,36 @@ public:
                 interpolatedColor.a = static_cast<uint8_t>(origin.a + (alphaToAdd * blend));
 
                 update(interpolatedColor);
+            }
+        }, completed, 0, pauseSensitive);
+    }
+
+    static H2DE_TimelineID lerp(H2DE_Engine* engine, const H2DE_Time& origin, const H2DE_Time& destination, H2DE_TimelineID duration, H2DE_Easing easing, const std::function<void(H2DE_Time)>& update, const std::function<void()>& completed, bool pauseSensitive = true) {
+        const uint8_t hrToAdd = destination.hours - origin.hours;
+        const uint8_t minToAdd = destination.minutes - origin.minutes;
+        const uint8_t secToAdd = destination.seconds - origin.seconds;
+        const uint16_t msToAdd = destination.milliseconds - origin.milliseconds;
+
+        return engine->createTimeline(duration, easing, [update, origin, hrToAdd, minToAdd, secToAdd, msToAdd](float blend) {
+            H2DE_Time interpolatedTime = H2DE_Time();
+            interpolatedTime.hours = static_cast<uint8_t>(origin.hours + (hrToAdd * blend));
+            interpolatedTime.minutes = static_cast<uint8_t>(origin.minutes + (minToAdd * blend));
+            interpolatedTime.seconds = static_cast<uint8_t>(origin.seconds + (secToAdd * blend));
+            interpolatedTime.milliseconds = static_cast<uint16_t>(origin.milliseconds + (msToAdd * blend));
+
+            update(interpolatedTime);
+
+        }, completed, 0, pauseSensitive);
+    }
+
+    template<typename H2DE_Lerp_T>
+    static H2DE_TimelineID lerp(H2DE_Engine* engine, H2DE_Lerp_T origin, H2DE_Lerp_T destination, H2DE_TimelineID duration, H2DE_Easing easing, const std::function<void(H2DE_Lerp_T)>& update, const std::function<void()>& completed, bool pauseSensitive = true) {
+        const H2DE_Lerp_T valueToAdd = destination - origin;
+
+        return engine->createTimeline(duration, easing, [update, origin, valueToAdd](float blend) {
+            if (update) {
+                H2DE_Lerp_T interpolatedValue = origin + (valueToAdd * blend);
+                update(interpolatedValue);
             }
         }, completed, 0, pauseSensitive);
     }