+++ src/engine/H2DE_timeline_manager.cpp
@@ -15,7 +15,7 @@ void H2DE_TimelineManager::update() {
     for (auto it = timelines.begin(); it != timelines.end(); ) {
         auto& [timelineId, timeline] = *it;
     
-        if (timeline.pauseSensitive && engine->isPaused()) {
+        if ((timeline.pauseSensitive && engine->isPaused()) || (timeline.paused)) {
             ++it;
             continue;
         }
@@ -27,7 +27,7 @@ void H2DE_TimelineManager::update() {
             blend = H2DE_Lerp(0.0f, 1.0f, blend, timeline.easing);
             timeline.update(blend);
         }
-    
+
         if (timeline.current >= timeline.duration) {
             if (timeline.completed) {
                 timeline.completed();
@@ -37,6 +37,7 @@ void H2DE_TimelineManager::update() {
             timeline.current = 0;
 
             if (timeline.loops == -1) {
+                ++it;
                 continue;
             }
 
@@ -53,17 +54,42 @@ void H2DE_TimelineManager::update() {
 // ACIOTNS
 H2DE_TimelineID H2DE_TimelineManager::create(H2DE_TimelineID duration, H2DE_Easing easing, const std::function<void(float)>& update, const std::function<void()>& completed, uint32_t loops, bool pauseSensitive) {
     H2DE_TimelineManager::H2DE_Timeline timeline = H2DE_TimelineManager::H2DE_Timeline();
-    timeline.duration = duration;
+    timeline.duration = std::max(duration, 1u);
     timeline.loops = loops;
     timeline.easing = easing;
     timeline.update = update;
     timeline.completed = completed;
     timeline.pauseSensitive = pauseSensitive;
+    timeline.paused = false;
 
     id++;
     timelines[id] = timeline;
 
     return id;
+}
+
+void H2DE_TimelineManager::pause(H2DE_TimelineID id) {
+    const auto it = timelines.find(id);
+
+    if (it != timelines.end()) {
+        it->second.paused = true;
+    }
+}
+
+void H2DE_TimelineManager::resume(H2DE_TimelineID id) {
+    const auto it = timelines.find(id);
+
+    if (it != timelines.end()) {
+        it->second.paused = false;
+    }
+}
+
+void H2DE_TimelineManager::togglePause(H2DE_TimelineID id) {
+    const auto it = timelines.find(id);
+
+    if (it != timelines.end()) {
+        it->second.paused = !it->second.paused;
+    }
 }
 
 void H2DE_TimelineManager::reset(H2DE_TimelineID id) {
@@ -91,4 +117,11 @@ void H2DE_TimelineManager::stop(H2DE_TimelineID id, bool callCompleted) {
 
         timelines.erase(it);
     }
+}
+
+// GETTER
+bool H2DE_TimelineManager::isPaused(H2DE_TimelineID id) const {
+    const auto it = timelines.find(id);
+
+    return (it != timelines.end()) ? it->second.paused : false;
 }